- changed definition of services.goodsla to double(5.2) (Alexei)

+++ create/mysql/schema.sql
@@ -29,7 +29,7 @@ CREATE TABLE services (
   algorithm		int(1)		DEFAULT '0' NOT NULL,
   triggerid		int(4),
   showsla		int(1)		DEFAULT '0' NOT NULL,
-  goodsla		double(3,2)	DEFAULT '99.9' NOT NULL,
+  goodsla		double(5,2)	DEFAULT '99.9' NOT NULL,
   sortorder		int(4)		DEFAULT '0' NOT NULL,
   PRIMARY KEY (serviceid)
 ) type=InnoDB;
